+++ src/components/groups/custom/requests/GroupRequested.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import RequestsSection from "./requests/RequestsSection";
+import RequestsSection from "./RequestsSection";
 import { useGroupRequests } from "@/src/hooks/groups/custom/requests/useGroupRequests";
 
 export default function GroupRequested({ groupID }: { groupID: string }) {
